all minters page, add titles
move asset and mint out of components
