+++ src/lib/components/Header.svelte
@@ -45,6 +45,9 @@
                 <span class="font-bold text-lg">Trending NFTs</span>
                 <span class="text-info">...</span>
                 <div class="card-actions">
+                  <a class="btn btn-primary btn-block" href="/explorer/minters">All Minters</a>
+                </div>
+                <div class="card-actions">
                   <a class="btn btn-primary btn-block" href="/market">View Market</a>
                 </div>
               </div>
